refactor: 입출력 뷰의 역할 명확하게 하도록 분리
